Add cli package
Issue #847
