+++ client_dev.py
@@ -1,7 +1,7 @@
 """
 Simple shim for running the client program during development.
 """
-import ga4gh.cli
+import ga4gh.cli.client as cli_client
 
 if __name__ == "__main__":
-    ga4gh.cli.client_main()
+    cli_client.client_main()
