+++ configtest_dev.py
@@ -1,7 +1,7 @@
 """
 Simple shim for running the configuration testing program during development.
 """
-import ga4gh.cli
+import ga4gh.cli.configtest as cli_configtest
 
 if __name__ == "__main__":
-    ga4gh.cli.configtest_main()
+    cli_configtest.configtest_main()
