+++ ga2vcf_dev.py
@@ -5,7 +5,7 @@ from __future__ import division
 from __future__ import print_function
 from __future__ import unicode_literals
 
-import ga4gh.cli
+import ga4gh.cli.ga2vcf as cli_ga2vcf
 
 if __name__ == "__main__":
-    ga4gh.cli.ga2vcf_main()
+    cli_ga2vcf.ga2vcf_main()
