+++ ga4gh/cli/configtest.py
@@ -1,0 +1,69 @@
+"""
+configtest cli
+"""
+from __future__ import division
+from __future__ import print_function
+from __future__ import unicode_literals
+
+import logging
+import unittest
+
+import ga4gh.cli as cli
+import ga4gh.configtest as configtest
+
+
+class SimplerResult(unittest.TestResult):
+    """
+    The TestResult class gives formatted tracebacks as error messages, which
+    is not what we want. Instead we just want the error message from the
+    err praram. Hence this subclass.
+    """
+    def addError(self, test, err):
+        self.errors.append((test,
+                            "{0}: {1}".format(err[0].__name__, err[1])))
+
+    def addFailure(self, test, err):
+        self.failures.append((test,
+                              "{0}: {1}".format(err[0].__name__, err[1])))
+
+
+def configtest_main(parser=None):
+    if parser is None:
+        parser = cli.createArgumentParser(
+            "GA4GH server configuration validator")
+    parser.add_argument(
+        "--config", "-c", default='DevelopmentConfig', type=str,
+        help="The configuration to use")
+    parser.add_argument(
+        "--config-file", "-f", type=str, default=None,
+        help="The configuration file to use")
+    cli.addVersionArgument(parser)
+
+    args = parser.parse_args()
+    configStr = 'ga4gh.serverconfig:{0}'.format(args.config)
+
+    configtest.TestConfig.configStr = configStr
+    configtest.TestConfig.configFile = args.config_file
+    configtest.TestConfig.configEnv = "GA4GH_CONFIGURATION"
+
+    loader = unittest.TestLoader()
+    tests = loader.loadTestsFromModule(configtest)
+    results = SimplerResult()
+    tests.run(results)
+
+    logging.basicConfig(level=logging.INFO)
+    log = logging.getLogger(__name__)
+    log.info('{0} Tests run. {1} errors, {2} failures, {3} skipped'.
+             format(results.testsRun,
+                    len(results.errors),
+                    len(results.failures),
+                    len(results.skipped)))
+    for result in results.errors:
+        if result is not None:
+            log.critical('Error: {0}: {1}'.format(result[0].id(), result[1]))
+    for result in results.failures:
+        if result is not None:
+            log.critical('Failure: {0}: {1}'.format(result[0].id(), result[1]))
+    for result in results.skipped:
+        if result is not None:
+            log.info('Skipped: {0}: {1}'.format(result[0].id(), result[1]))
