+++ repo_dev.py
@@ -1,7 +1,7 @@
 """
 Simple shim for running the repo program during development.
 """
-import ga4gh.cli
+import ga4gh.cli.repomanager as cli_repomanager
 
 if __name__ == "__main__":
-    ga4gh.cli.repo_main()
+    cli_repomanager.repo_main()
