+++ server_dev.py
@@ -1,7 +1,7 @@
 """
 Simple shim for running the server program during development.
 """
-import ga4gh.cli
+import ga4gh.cli.server as cli_server
 
 if __name__ == "__main__":
-    ga4gh.cli.server_main()
+    cli_server.server_main()
