+++ setup.py
@@ -34,12 +34,12 @@ setup(
     url="https://github.com/ga4gh/server",
     entry_points={
         'console_scripts': [
-            'ga4gh_client=ga4gh.cli:client_main',
-            'ga4gh_configtest=ga4gh.cli:configtest_main',
-            'ga4gh_server=ga4gh.cli:server_main',
-            'ga2vcf=ga4gh.cli:ga2vcf_main',
-            'ga2sam=ga4gh.cli:ga2sam_main',
-            'ga4gh_repo=ga4gh.cli:repo_main',
+            'ga4gh_client=ga4gh.cli.client:client_main',
+            'ga4gh_configtest=ga4gh.cli.configtest:configtest_main',
+            'ga4gh_server=ga4gh.cli.server:server_main',
+            'ga2vcf=ga4gh.cli.ga2vcf:ga2vcf_main',
+            'ga2sam=ga4gh.cli.ga2sam:ga2sam_main',
+            'ga4gh_repo=ga4gh.cli.repomanager:repo_main',
         ]
     },
     classifiers=[
