+++ tests/end_to_end/test_client_json.py
@@ -13,7 +13,7 @@ import unittest
 
 import ga4gh.client as client
 import ga4gh.backend as backend
-import ga4gh.cli as cli
+import ga4gh.cli.client as cli_client
 import ga4gh.protocol as protocol
 import ga4gh.datarepo as datarepo
 import tests.utils as utils
@@ -42,7 +42,7 @@ class TestClientOutput(unittest.TestCase):
         clientCommand = "{} {} {} -O {}".format(
             command, self._dataUrl, arguments, outputFormat)
         stdout, stderr = utils.captureOutput(
-            cli.client_main, clientCommand.split())
+            cli_client.client_main, clientCommand.split())
         self.assertEqual(len(stderr), 0)
         return stdout
